requirements.txt
Flask==0.10.1
Jinja2==2.7.3
MarkupSafe==0.23
Pillow==2.8.1
Werkzeug==0.10.4
argparse==1.2.1
blinker==1.3
itsdangerous==0.24
requests==2.7.0
rollbar==0.9.13
wsgiref==0.1.2
